Add support for root Path Prefix
When a call using primer client is done for `GET /v1/test`,
with the current implementation the call gets translated to:
`http://host:port/v1/test`

Some API gateways use a rootPathPrefix to decide which backend to send
a request. If we need to translate this call to a call accepted by gateway,
it should become:
`http://host:port/apis/backend/v1/test.`
where `apis/backend` is the `rootPathPrefix` to reach the backend

It would be better to have this capability in Primer Client.

Also added support for secure endpoints.

Signed-off-by: Pranith Kumar K <pranith.karampuri@phonepe.com>